+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Menu drawer not working well on small screens
+
 ## [0.2.0] - 2025-02-21
 
 ### Added